refactor: move general ledger state into providers/generalLedger
The journal route store joins the three ledger contexts, which sit beside
LedgerDateStore from the earlier commit. ChartOfAccountsContext could argue for
customAccounts, but it is provided inline by features/generalLedger's
ChartOfAccounts and Journal components, so generalLedger is where a reader
following the render tree will look for it.
