+++ src/components/features/generalLedger/ChartOfAccounts/ChartOfAccounts.tsx
@@ -3,10 +3,10 @@ import { type ReactNode, useContext } from 'react'
 import { useChartOfAccounts } from '@hooks/legacy/useChartOfAccounts'
 import { useLedgerAccounts } from '@hooks/legacy/useLedgerAccounts'
 import { useElementViewSize } from '@hooks/utils/size/useElementViewSize'
+import { ChartOfAccountsContext } from '@providers/generalLedger/ChartOfAccountsContext/ChartOfAccountsContext'
+import { LedgerAccountsContext } from '@providers/generalLedger/LedgerAccountsContext/LedgerAccountsContext'
 import { LedgerDateStoreProvider } from '@providers/generalLedger/LedgerDateStore/LedgerDateStoreProvider'
 import { InAppLinkProvider, type LinkingMetadata } from '@providers/global/InAppLink/InAppLinkContext'
-import { ChartOfAccountsContext } from '@contexts/ChartOfAccountsContext/ChartOfAccountsContext'
-import { LedgerAccountsContext } from '@contexts/LedgerAccountsContext/LedgerAccountsContext'
 import { Loader } from '@ui/Loader/Loader'
 import { Container } from '@blocks/Layout/Container/Container'
 import { type ChartOfAccountsTableStringOverrides, ChartOfAccountsTableWithPanel } from '@features/generalLedger/ChartOfAccountsTableWithPanel/ChartOfAccountsTableWithPanel'
