+++ src/components/features/generalLedger/ChartOfAccountsForm/ChartOfAccountsForm.tsx
@@ -6,8 +6,8 @@ import { useTranslation } from 'react-i18next'
 import { type LedgerAccountType, type LedgerEntryDirection, type NestedLedgerAccountType } from '@schemas/generalLedger/ledgerAccount'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
 import { UpsertMode } from '@hooks/utils/swr/createUpsertHook'
+import { ChartOfAccountsContext } from '@providers/generalLedger/ChartOfAccountsContext/ChartOfAccountsContext'
 import { useLayerContext } from '@providers/global/LayerContext/LayerContext'
-import { ChartOfAccountsContext } from '@contexts/ChartOfAccountsContext/ChartOfAccountsContext'
 import { Button } from '@ui/Button/Button'
 import { CloseButton } from '@ui/Button/CloseButton'
 import { DataState, DataStateStatus } from '@ui/DataState/DataState'
