+++ src/components/features/generalLedger/Journal/Journal.tsx
@@ -3,11 +3,11 @@ import { type ReactNode } from 'react'
 import { useChartOfAccounts } from '@hooks/legacy/useChartOfAccounts'
 import { useJournal } from '@hooks/legacy/useJournal'
 import { useElementViewSize } from '@hooks/utils/size/useElementViewSize'
+import { ChartOfAccountsContext } from '@providers/generalLedger/ChartOfAccountsContext/ChartOfAccountsContext'
+import { JournalContext } from '@providers/generalLedger/JournalContext/JournalContext'
+import { JournalRoute, JournalStoreProvider, useJournalRouteState } from '@providers/generalLedger/JournalStore/JournalStoreProvider'
 import { LedgerDateStoreProvider } from '@providers/generalLedger/LedgerDateStore/LedgerDateStoreProvider'
 import { InAppLinkProvider, type LinkingMetadata } from '@providers/global/InAppLink/InAppLinkContext'
-import { JournalRoute, JournalStoreProvider, useJournalRouteState } from '@providers/JournalStore/JournalStoreProvider'
-import { ChartOfAccountsContext } from '@contexts/ChartOfAccountsContext/ChartOfAccountsContext'
-import { JournalContext } from '@contexts/JournalContext/JournalContext'
 import { Loader } from '@ui/Loader/Loader'
 import { Container } from '@blocks/Layout/Container/Container'
 import { JournalEntryDrawer } from '@features/generalLedger/JournalEntryDrawer/JournalEntryDrawer'
