+++ src/components/features/generalLedger/JournalEntryForm/JournalEntryForm.tsx
@@ -9,7 +9,7 @@ import { flattenValidationErrors } from '@utils/form'
 import { usePreloadCustomers } from '@api/businesses/[business-id]/customers/get'
 import { UpsertJournalEntryMode } from '@api/businesses/[business-id]/ledger/journal-entries/post'
 import { usePreloadVendors } from '@api/businesses/[business-id]/vendors/get'
-import { useJournalNavigation } from '@providers/JournalStore/JournalStoreProvider'
+import { useJournalNavigation } from '@providers/generalLedger/JournalStore/JournalStoreProvider'
 import { DataState, DataStateStatus } from '@ui/DataState/DataState'
 import { Form } from '@ui/Form/Form'
 import { Separator } from '@ui/Separator/Separator'
