+++ src/components/features/generalLedger/JournalTable/JournalTable.tsx
@@ -10,8 +10,8 @@ import { entryNumber, sumLineItemAmountsByDirection } from '@utils/journal'
 import { JOURNAL_PAGE_SIZE } from '@hooks/legacy/useJournal'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { usePaginatedList } from '@hooks/utils/pagination/usePaginatedList'
+import { JournalContext } from '@providers/generalLedger/JournalContext/JournalContext'
 import { useLayerContext } from '@providers/global/LayerContext/LayerContext'
-import { JournalContext } from '@contexts/JournalContext/JournalContext'
 import { DataState, DataStateStatus } from '@ui/DataState/DataState'
 import { Pagination } from '@ui/Pagination/Pagination'
 import { MoneySpan } from '@ui/Typography/MoneySpan'
