+++ src/components/features/generalLedger/JournalTableWithPanel/JournalTableWithPanel.tsx
@@ -3,9 +3,9 @@ import { CirclePlus } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
-import { useJournalNavigation } from '@providers/JournalStore/JournalStoreProvider'
+import { JournalContext } from '@providers/generalLedger/JournalContext/JournalContext'
+import { useJournalNavigation } from '@providers/generalLedger/JournalStore/JournalStoreProvider'
 import { useBookkeepingStatusContext } from '@contexts/BookkeepingStatusContext/BookkeepingStatusContext'
-import { JournalContext } from '@contexts/JournalContext/JournalContext'
 import { Button } from '@ui/Button/Button'
 import { Heading } from '@ui/Typography/Heading'
 import { LedgerDateRangeSelection } from '@blocks/DatePickers/DateSelection/LedgerDateRangeSelection'
